create filters file
