+++ .eleventy.js
@@ -15,26 +15,11 @@ module.exports = function (eleventyConfig) {
 
   require('./utils/imgix')(eleventyConfig);
   require('./utils/shortcodes')(eleventyConfig);
+  require('./utils/filters')(eleventyConfig);
 
   eleventyConfig.setDataDeepMerge(true);
 
   eleventyConfig.addLayoutAlias('post', 'layouts/post.njk');
-
-  eleventyConfig.addFilter(
-    'dateForDisplay',
-    (dateString, format = 'fff', opts = {}) => {
-      const resolvedOptions = {
-        ...opts,
-      };
-      let d = dateString;
-      if (typeof d === 'object' && d.toISOString) {
-        d = d.toISOString();
-      }
-      return DateTime.fromISO(d)
-        .setZone('America/New_York')
-        .toFormat(format, resolvedOptions);
-    }
-  );
 
   eleventyConfig.addCollection('homePageBlocksSmall', function (collectionApi) {
     return collectionApi
@@ -60,47 +45,11 @@ module.exports = function (eleventyConfig) {
       );
   });
 
-  eleventyConfig.addFilter('toLocaleString', (number, locale = 'en-US') => {
-    const parsed = parseFloat(number);
-    return isNaN(parsed) ? number : parsed.toLocaleString(locale);
-  });
-
-  // Get the first `n` elements of a collection.
-  eleventyConfig.addFilter('head', (array, n) => {
-    if (n < 0) {
-      return array.slice(n);
-    }
-
-    return array.slice(0, n);
-  });
-
   eleventyConfig.addPassthroughCopy('src/img');
   eleventyConfig.addPassthroughCopy('src/service-worker.js');
   eleventyConfig.addPassthroughCopy('src/assets');
   eleventyConfig.addPassthroughCopy('src/css');
   // eleventyConfig.addPassthroughCopy('CODE_OF_CONDUCT.md');
-
-  eleventyConfig.addFilter('assetPath', function (value) {
-    if (process.env.ELEVENTY_ENV === 'production') {
-      const manifestPath = 'assets.json';
-      const manifest = JSON.parse(fs.readFileSync(manifestPath));
-      return manifest[value] ? manifest[value] : value;
-    }
-    return value;
-  });
-
-  eleventyConfig.addFilter('qualifiedUrl', function (path) {
-    if (!process.env.NETLIFY) {
-      return path;
-    }
-
-    const baseUrl =
-      process.env.CONTEXT === 'production'
-        ? process.env.URL
-        : process.env.DEPLOY_PRIME_URL;
-
-    return baseUrl ? baseUrl + path : path;
-  });
 
   const markdownItAttrs = require('markdown-it-attrs');
